Add cruid by user id

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { Task } from './task.entity';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -12,8 +13,13 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository
     ) {}
 
-  async getById(id: number): Promise<Task>  {
-    const found = await this.taskRepository.findOne(id);
+  async getById(id: number, user: User): Promise<Task>  {
+    const found = await this.taskRepository.findOne({
+      where: {
+        id,
+        userId: user.id
+      }
+    });
 
     if (!found) {
       throw new NotFoundException(`Task with id: '${id}' not found`);
@@ -22,24 +28,24 @@ export class TasksService {
     return found
   }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
 
-  create(taskData: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(taskData);
+  create(taskData: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(taskData, user);
   }
 
-  async delete(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async delete(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
     if (result.affected === 0) {
       throw new NotFoundException(`Task with id: '${id}' not found`);
     }
   }
 
-  async updateStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getById(id);
+  async updateStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getById(id, user);
     task.status = status;
     await task.save();
     return task;
